refactor(SBOMER-218): moving SyftImageController to top-levek 'generator' package, rest will follow
